Rename 'Invocations' component to 'InvocationsList' component to satify multi-word component name
