+++ client/src/components/Workflow/StoredWorkflowInvocations.vue
@@ -1,7 +1,7 @@
 <template>
     <CurrentUser v-slot="{ user }">
         <StoredWorkflowDetailsProvider v-slot="{ result: item, loading }" :stored-workflow-id="storedWorkflowId">
-            <Invocations
+            <InvocationsList
                 v-if="!loading && user.id"
                 :user-id="user.id"
                 :stored-workflow-id="item.id"
@@ -10,14 +10,14 @@
     </CurrentUser>
 </template>
 <script>
-import Invocations from "components/Workflow/Invocations";
+import InvocationsList from "components/Workflow/InvocationsList";
 import CurrentUser from "components/providers/CurrentUser";
 import { StoredWorkflowDetailsProvider } from "components/providers/StoredWorkflowsProvider";
 
 export default {
     components: {
         CurrentUser,
-        Invocations,
+        InvocationsList,
         StoredWorkflowDetailsProvider,
     },
     props: {
