+++ client/src/components/Workflow/UserInvocations.vue
@@ -1,15 +1,15 @@
 <template>
     <!-- Bit of an odd component, but the API returns all invocations for admins if no user id is included in request-->
-    <CurrentUser v-slot="{ user }"> <Invocations v-if="user.id" :user-id="user.id" /> </CurrentUser>
+    <CurrentUser v-slot="{ user }"> <InvocationsList v-if="user.id" :user-id="user.id" /> </CurrentUser>
 </template>
 <script>
-import Invocations from "components/Workflow/Invocations";
+import InvocationsList from "components/Workflow/InvocationsList";
 import CurrentUser from "components/providers/CurrentUser";
 
 export default {
     components: {
         CurrentUser,
-        Invocations,
+        InvocationsList,
     },
 };
 </script>
