+++ client/src/components/admin/ActiveInvocations.vue
@@ -1,17 +1,17 @@
 <template>
-    <invocations
+    <invocations-list
         header-message="Workflow invocations that are still being scheduled are displayed on this page."
         no-invocations-message="There are no scheduling workflow invocations to show currently."
         :owner-grid="false">
-    </invocations>
+    </invocations-list>
 </template>
 
 <script>
-import Invocations from "../Workflow/Invocations";
+import InvocationsList from "../Workflow/InvocationsList";
 
 export default {
     components: {
-        Invocations,
+        InvocationsList,
     },
 };
 </script>
